refactor(hs): add new helper for inputs
