+++ packages/headless-styles/src/components/Textarea/shared.ts
@@ -1,4 +1,3 @@
-import { createA11yProps } from '../../utils/helpers'
 import { getDefaultInputFieldOptions } from '../shared/defaultOptions'
 import type { TextareaOptions, TextareaResize } from './types'
 
@@ -13,22 +12,5 @@ export function getDefaultTextareaOptions(options?: TextareaOptions) {
 export function getTextareaClasses(resize: TextareaResize) {
   return {
     resizeClass: `pando_${resize}Textarea` as const,
-  }
-}
-
-export function createTextareaOptions(options: TextareaOptions) {
-  const { describedBy } = options
-  const a11yProps = createA11yProps(options)
-  const describedByProps = describedBy && {
-    ['aria-describedby']: describedBy,
-  }
-
-  return {
-    ...a11yProps,
-    ...describedByProps,
-    id: options.id,
-    name: options.name,
-    placeholder: options.placeholder,
-    value: options.value,
   }
 }
